requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
addict==1.0.0
astroid==1.4.9            # via pylint
backports.functools-lru-cache==1.3  # via pylint
boto3==1.4.7
botocore==1.7.5           # via boto3, s3transfer
clip.py==0.3.2
configparser==3.5.0       # via pylint
docutils==0.14            # via botocore
functools32==3.2.3-2
futures==3.1.1            # via s3transfer
isort==4.2.5              # via pylint
jmespath==0.9.3           # via boto3, botocore
lazy-object-proxy==1.2.2  # via astroid
logtool==0.3.post16
mccabe==0.5.3             # via pylint
progress==1.2
pylint==1.6.4             # via setuptools-lint
python-dateutil==2.6.1    # via botocore
retryp==0.3.post9
s3transfer==0.1.11        # via boto3
setuptools-lint==0.5.2
six==1.10.0               # via astroid, logtool, pylint, python-dateutil
termcolor==1.1.0
wrapt==1.10.8             # via astroid, logtool, retryp
